+++ src/functions/game-result.js
